core/pyproject.toml
[tool.poetry]
name = "small-django"
version = "0.1.0"
description = "Simple project with 3-4 models"
authors = ["radthenone <krzyzak21@gmail.com>"]
readme = "README.md"
packages = [{include = "small_django"}]

[tool.poetry.dependencies]
python = "3.10.7"
django-environ = "^0.9.0"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
drf-spectacular = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
